+++ src/memory/ReproCartridgeV2.cc
@@ -1,6 +1,7 @@
 #include "ReproCartridgeV2.hh"
 #include "DummyAY8910Periphery.hh"
 #include "MSXCPUInterface.hh"
+#include "narrow.hh"
 #include "serialize.hh"
 #include <array>
 
@@ -324,9 +325,9 @@ void ReproCartridgeV2::setVolume(EmuTime::param time, byte value)
 	// EE (7-6) = PSG#10. So values 0-3
 	// III(5-3) = PSG#A0. So values 0-7
 	// SSS(2-0) = SCC.    So values 0-7
-	scc    .setSoftwareVolume(((volumeReg >> 0) & 7) / 2.0f, time);
-	psg0xA0.setSoftwareVolume(((volumeReg >> 3) & 7) / 2.0f, time);
-	psg0x10.setSoftwareVolume(((volumeReg >> 6) & 3) / 2.0f, time);
+	scc    .setSoftwareVolume(narrow<float>((volumeReg >> 0) & 7) / 2.0f, time);
+	psg0xA0.setSoftwareVolume(narrow<float>((volumeReg >> 3) & 7) / 2.0f, time);
+	psg0x10.setSoftwareVolume(narrow<float>((volumeReg >> 6) & 3) / 2.0f, time);
 }
 
 template<typename Archive>
